moved xstate machine def form to a new file
